+++ src/startup-routines/find_param_fixform.f
@@ -62,7 +62,7 @@ C
       USE EIRMOD_COMNNL, ONLY: NPRNLI
       USE EIRMOD_COMPRT, ONLY: IUNIN, IUNOUT
       USE EIRMOD_CLOGAU, ONLY: NLWRMSH, NLSPCSCL, NLSPCSCL_ON,
-     .                         NLSOLEDGE
+     .                         NLSOLEDGE, NLSPCSPT
       USE EIRMOD_CTRCEI, ONLY: TRCAMD, TRCINT, NVOLPR, NSURPR,
      .                         EIRENE_ALLOC_CTRCEI
       USE EIRMOD_CINIT, ONLY: CASENAME, DBFNAME, DBHANDLE, NDBNAMES,
@@ -206,11 +206,12 @@ C                                    OR =9  (FULL A&M STORAGE MODE, =DEFAULT)
       IF (NSTORAM < 9) NSTORAM = 0
       NOPTM1 = MAX(NOPTM1,1)
 C
-      call fix_logical_input(zeile,19)
+      call fix_logical_input(zeile,20)
       READ (ZEILE,6665) NLSCL,NLTEST,NLANA,NLDRFT,NLCRR,
      .                  NLERG,NLIDENT,NLONE,NLMOVIE,NLDFST,
      .                  NLRANMAR,NLCASCAD,NLOCTREE,NLWRMSH,NEXVS,
-     .                  NLTRIMESH,NLSPCSCL,NLSPCSCL_ON,NLSOLEDGE
+     .                  NLTRIMESH,NLSPCSCL,NLSPCSCL_ON,NLSOLEDGE,
+     .                  NLSPCSPT
 
       CALL EIRENE_INIT_CINIT